Code review: core.

+++ FluidEngine/src/FluidEngine/Core/Events/ApplicationEvent.h
@@ -10,14 +10,14 @@ namespace fe {
 	class WindowResizeEvent : public Event
 	{
 	public:
-		WindowResizeEvent(unsigned int width, unsigned int height)
+		WindowResizeEvent(uint16_t width, uint16_t height)
 			: m_Width(width), m_Height(height) {}
 
-		inline unsigned int GetWidth() const { 
+		inline uint16_t GetWidth() const { 
 			return m_Width; 
 		}
 
-		inline unsigned int GetHeight() const {
+		inline uint16_t GetHeight() const {
 			return m_Height; 
 		}
 
@@ -31,7 +31,7 @@ namespace fe {
 		EVENT_CLASS_TYPE(WindowResize)
 		EVENT_CLASS_CATEGORY(EventCategoryApplication)
 	private:
-		unsigned int m_Width, m_Height;
+		uint16_t m_Width, m_Height;
 	};
 
 	/// <summary>
